+++ ost/src/server.cpp
@@ -27,13 +27,15 @@ namespace ostbackend {
 
 Server::Server(
     unsigned int queue_size, unsigned int write_throttle_limit,
-    const std::string& addr, unsigned int port, const char* location
+    uint64_t write_backlog_limit, const std::string& addr, unsigned int port,
+    const char* location
 ) :
     _queue(nullptr),
     _fd(-1),
     _locations(rawstd::URI::uriv(location)),
     _accept_event(nullptr),
-    _write_throttle_limit(write_throttle_limit) {
+    _write_throttle_limit(write_throttle_limit),
+    _write_backlog_limit(write_backlog_limit) {
 
     int res = rawstor_initialize(nullptr);
     if (res < 0) {